requirements.txt
Flask
pandas
Pillow
google-cloud-vision 